# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Oxymed Medikal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /product-categories:
    get:
      operationId: listProductCategories
      tags: [products]
      summary: List product categories
      responses:
        "200":
          description: List of product categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductCategory"
    post:
      operationId: createProductCategory
      tags: [products]
      summary: Create product category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductCategory"
  /product-categories/{id}:
    patch:
      operationId: updateProductCategory
      tags: [products]
      summary: Update product category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductCategory"
    delete:
      operationId: deleteProductCategory
      tags: [products]
      summary: Delete product category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PRODUCTS
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: published
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # NEWS
...
  /references:
    get:
      operationId: listReferences
      tags: [references]
      summary: List reference projects
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of references
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceListResponse"
    post:
      operationId: createReference
      tags: [references]
      summary: Create reference project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceItem"
  /references/{id}:
    patch:
      operationId: updateReference
      tags: [references]
      summary: Update reference project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferenceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceItem"
    delete:
      operationId: deleteReference
      tags: [references]
      summary: Delete reference project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # QUOTE REQUESTS
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List quote requests (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quote requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteListResponse"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Submit a quote request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Quote request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteItem"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get quote request by id (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteItem"
    patch:
      operationId: updateQuoteStatus
      tags: [quotes]
      summary: Update quote request status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteItem"

  # CORPORATE
...
  /settings/{settingKey}:
    put:
      operationId: upsertSetting
      tags: [settings]
      summary: Create or update a setting
      security:
        - bearerAuth: []
      parameters:
        - name: settingKey
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingInput"
      responses:
        "200":
          description: Setting upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"

  # MEDIA
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AUTH
...
    ProductCategory:
      type: object
      required: [id, name, slug, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductCategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
    ProductCategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer

    # PRODUCTS
    ProductSpec:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: string
    Product:
      type: object
      required: [id, title, published, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        categoryId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        sortOrder:
          type: integer
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductInput:
      type: object
      required: [title]
      properties:
        categoryId:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        sortOrder:
          type: integer
        published:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        sortOrder:
          type: integer
        published:
          type: boolean
    ProductListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer

    # NEWS
...
    ReferenceItem:
      type: object
      required: [id, title, projectType, category, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        projectType:
          type: string
        capacity:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    ReferenceInput:
      type: object
      required: [title, projectType]
      properties:
        title:
          type: string
        projectType:
          type: string
        capacity:
          type: string
        city:
          type: string
        imageUrl:
          type: string
        category:
          type: string
    ReferenceUpdate:
      type: object
      properties:
        title:
          type: string
        projectType:
          type: string
        capacity:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
    ReferenceListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReferenceItem"
        total:
          type: integer

    # QUOTES
    QuoteItem:
      type: object
      required: [id, fullName, email, phone, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        projectType:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        applicationArea:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    QuoteInput:
      type: object
      required: [fullName, email, phone]
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        jobTitle:
          type: string
        projectType:
          type: string
        city:
          type: string
        applicationArea:
          type: string
        notes:
          type: string
    QuoteStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    QuoteListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        total:
          type: integer

    # CORPORATE
...
    Setting:
      type: object
      required: [id, settingKey, settingValue]
      properties:
        id:
          type: integer
        settingKey:
          type: string
        settingValue:
          type: string
    SettingInput:
      type: object
      required: [settingValue]
      properties:
        settingValue:
          type: string
...