# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oxymed Medikal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpec } from './productSpec';

export interface Product {
  id: number;
  /** @nullable */
  categoryId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  specs?: ProductSpec[];
  sortOrder: number;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}