# lib/api-zod/src/generated/types/productCategoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oxymed Medikal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductCategoryInput {
  name: string;
  slug: string;
  description?: string;
  sortOrder?: number;
}