# lib/api-zod/src/generated/types/productCategoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oxymed Medikal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductCategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
}