# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oxymed Medikal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpec } from './productSpec';

export interface ProductInput {
  categoryId?: number;
  title: string;
  description?: string;
  imageUrl?: string;
  specs?: ProductSpec[];
  sortOrder?: number;
  published?: boolean;
}