# lib/api-zod/src/generated/types/referenceListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oxymed Medikal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReferenceItem } from './referenceItem';

export interface ReferenceListResponse {
  items: ReferenceItem[];
  total: number;
}